requirements.txt
numpy
torch
torchvision
python-dotenv
seaborn
matplotlib
tqdm